refactor: 모임마다 조회하던 참여자수, 찜을 한 번에 조회하도록 변경

+++ backend/src/main/java/mouda/backend/moim/domain/MoimOverview.java
@@ -9,9 +9,9 @@ public class MoimOverview {
 	private final int currentPeople;
 	private final boolean isZzimed;
 
-	public MoimOverview(Moim moim, int currentPeople, boolean isZzimed) {
+	public MoimOverview(Moim moim, long currentPeople, boolean isZzimed) {
 		this.moim = moim;
-		this.currentPeople = currentPeople;
+		this.currentPeople = (int)currentPeople;
 		this.isZzimed = isZzimed;
 	}
 }
